requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dnspython==2.1.0
email-validator==1.1.2
emoji==1.2.0
fastapi==0.63.0
future==0.18.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kaleido==0.2.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
nltk==3.6.2
numexpr==2.7.3
numpy==1.20.2
oauthlib==3.1.0
orca==1.6
orjson==3.5.2
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
promise==2.3
psutil==5.8.0
pydantic==1.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-multipart==0.0.5
python-twitter==3.5
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
Rx==1.6.1
six==1.15.0
starlette==0.13.6
tables==3.6.1
text2emotion==0.0.5
toolz==0.11.1
tqdm==4.60.0
typing-extensions==3.10.0.0
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
webencodings==0.5.1
websockets==8.1
